Make it easy to create liquid asset network definitions
