+++ NBitcoin.Altcoins/AltcoinNetworkSets.cs
@@ -27,6 +27,7 @@ namespace NBitcoin.Altcoins
 		public static Colossus Colossus { get; } = Colossus.Instance;
 		public static GoByte GoByte { get; } = GoByte.Instance;
 		public static Liquid Liquid { get; } = Liquid.Instance;
+		public static LiquidTether LiquidTether { get; } = LiquidTether.Instance;
 		public static Koto Koto { get; } = Koto.Instance;
 		public static Bitcoinplus Bitcoinplus { get; } = Bitcoinplus.Instance;
 		public static Chaincoin Chaincoin { get; } = Chaincoin.Instance;
@@ -54,6 +55,7 @@ namespace NBitcoin.Altcoins
 			yield return GoByte;
 			yield return Stratis;
 			yield return Liquid;
+			yield return LiquidTether;
 			yield return Koto;
 			yield return Bitcoinplus;
 			yield return Chaincoin;
